temp
There is a bug while combining, some entries are missed.
A zero entry should be added so that while combining it can
fallback to the zero entry instead of missing the entry.
This branch is still not complete, due to an error in the code due to
incorrect matching logic
aavePnlResponse.result.data is n length
however with addNullEntry it's n+1 length.

fix this later